third_party/blink/web_tests/flag-specific/highdpi/transforms/selection-bounds-in-transformed-view-expected.txt

This is a testharness.js-based test.
FAIL Test scrolled into view with prefer compositing enabled assert_approx_equals: expected 167 +/- 1 but got 163.6666717529297
FAIL Test scrolled into view with prefer compositing disabled assert_approx_equals: expected 167 +/- 1 but got 163.6666717529297
Harness: the test ran to completion.
